refactor: to fix filter menu state on new URL with searchparam
